+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - "Don't close the window" message when loading course files.
 
+### Changed
+- Updated translations api to chrome.i18n
+
 ### Fixed
 - button glow effect not using theme colors.
 - Icon color not updating correctly on browser start.